Code quality: history and herder cleanup batch
- Replace inline % 64 == 63 with is_checkpoint_ledger() (Closes #904)
- Remove dangerous CHECKPOINT_FREQUENCY const alias (Closes #900)
- Unify compare_transactions/compare_results with generic compare_entries<T> (Closes #897)
- Narrow herder persistence.rs visibility to pub(crate) (Closes #893)
- Extract named constants and unify LEDGER_VALIDITY_BRACKET in herder (Closes #888)

+++ crates/herder/src/sync_recovery.rs
@@ -65,7 +65,7 @@ pub const OUT_OF_SYNC_RECOVERY_INTERVAL: Duration = Duration::from_secs(10);
 
 /// Maximum number of slots to keep ahead when purging during recovery.
 /// Spec: HERDER_SPEC §16 — LEDGER_VALIDITY_BRACKET = 100.
-pub const LEDGER_VALIDITY_BRACKET: u32 = 100;
+pub const LEDGER_VALIDITY_BRACKET: u64 = 100;
 
 /// Commands sent to the sync recovery manager.
 #[derive(Debug)]
